My first SimulatedAnnealing.cpp
This algorithm is based on the process of annealing, which is a way of slowly cooling a metal to make it stronger. It works by starting with a random solution and then gradually making changes to it. If the changes improve the solution, they are kept. If the changes make the solution worse, they are sometimes kept anyway, with a probability that decreases as the algorithm progresses.